+++ src/main/ts/CreateToolSidebar.tsx
@@ -1,4 +1,6 @@
-namespace contoureditor {
+// namespace contoureditor {
+import React, { Component } from "react";
+import { BaseToolSidebar } from "./BaseToolSidebar";
 
     export class CreateToolSidebar extends BaseToolSidebar{
         public render() {
@@ -13,4 +15,4 @@ namespace contoureditor {
             );
         }
     }
-}
+// }