+++ src/main/ts/index.tsx
@@ -1,11 +1,16 @@
-///<reference path="3rdparty/react.development.js" />
-///<reference path="3rdparty/react-dom.development.js" />
-///<reference path="d.ts/react-dom/client.d.ts"/>
-///<reference path="App.tsx"/>
-///<reference path="model/Node.ts"/>
+// /<reference path="3rdparty/react.development.js" />
+// /<reference path="3rdparty/react-dom.development.js" />
+// /<reference path="d.ts/react-dom/client.d.ts"/>
+// /<reference path="App.tsx"/>
+// /<reference path="model/Node.ts"/>
+import React, { Component } from "react";
+import ReactDOM from 'react-dom/client';
+import { App } from "./App";
 
-namespace contoureditor {
-    document.addEventListener('DOMContentLoaded', function(){
+// namespace contoureditor {
+// export default
+// alert(1);
+//     document.addEventListener('DOMContentLoaded', function(){
         var wrapper = document.getElementById("contoureditor-wrapper");
         if( wrapper ) {
 
@@ -16,5 +21,5 @@ namespace contoureditor {
                 </React.StrictMode>
             );
         };
-    });
-}
+//     });
+// }